Extract ConvertV1Profiles out of ConfigBuilder
