+++ packages/imperative/src/config/index.ts
@@ -11,23 +11,24 @@
 
 export * from "./src/Config";
 export * from "./src/ConfigAutoStore";
+export * from "./src/ConfigBuilder";
 export * from "./src/ConfigConstants";
 export * from "./src/ConfigSchema";
-export * from "./src/ConfigBuilder";
 export * from "./src/ConfigUtils";
+export * from "./src/ConvertV1Profiles";
 export * from "./src/ProfileCredentials";
 export * from "./src/ProfileInfo";
 export * from "./src/ProfInfoErr";
 export * from "./src/doc/IConfig";
 export * from "./src/doc/IConfigAutoStoreOpts";
 export * from "./src/doc/IConfigBuilderOpts";
-export * from "./src/doc/IConfigConvert";
 export * from "./src/doc/IConfigLayer";
 export * from "./src/doc/IConfigOpts";
 export * from "./src/doc/IConfigProfile";
 export * from "./src/doc/IConfigSchema";
 export * from "./src/doc/IConfigSecure";
 export * from "./src/doc/IConfigVault";
+export * from "./src/doc/IConvertV1Profiles";
 export * from "./src/doc/IProfArgAttrs";
 export * from "./src/doc/IProfAttrs";
 export * from "./src/doc/IProfInfoUpdatePropOpts";
